requirements.txt
streamlit
python-dotenv
langchain
langchain_community
boto3
langchain-aws
pyyaml
opensearch-py
tqdm